split methods module
